Add support for creating a fat JAR using the shadow plugin.
Move the jEdit dependencies to the shadow configuration to simulate the
equivalent of a Maven "provided" configuration.  That is, we don't want
these dependencies included in the fat JAR.

This solution doesn't work exactly as I would like.  Per issue #166,
shadow does not work with the commonly-used nebula plugin
configurations.  Therefore, instead of a true provided configuration,
we're forced to use the shadow configuration, as suggested is issue
#142.  However, the drawback of this approach is that the JARs included
in the shadow configuration are added to the manifest's Class-Path
property, which is undesirable.

I'm committing this change simply to record how I achieved
not-quite-perfect fat JAR support using the shadow plugin.  A
forthcoming change will probably revert this and use the nebula plugin
with a roll-your-own fat JAR task.