Add Git command for comparing two non-repo files
Note that the semantics of the `git diff` command required changing the
assumptions of the `IGitRunner` interface.  Previously, we assumed any
nonzero exit code from the Git process indicated an error.  However,
this is not true for `git diff` (and several other commands, as I've
since discovered).  For example, `git diff` returns zero if the files
are identical and one if they are different.

Per [1], the most reliable way to determine if Git exited with an error
is to examine the content of the standard error stream.  If it is not
empty, one can assume an error occurred.  Otherwise, regardless of the
exit code, one should assume the operation was successful.

[1] http://www.gelato.unsw.edu.au/archives/git/0701/36671.html