Add the REMOVED_ABOVE_AND_BELOW dirty mark type
It is possible for the adjacent lines both above and below the target
line to be removed simultaneously.  If the target line has not been
changed, it must be able to display removed markers for both lines above
and below it.

The `PatchAnalyzer` unit tests have been modified in this change to no
longer manually create the `Patch` instances.  Instead, the associated
`difflib` library is used to generate the patches for two reasons: 1) to
make the tests easier to read, and 2) it is anticipated that the plugin
will be diffing in memory rather than delegating to the `git diff`
command; in that case, `difflib` is the library we will be using, so we
might as well ensure we are working on patches produced using that
algorithm.