Add task for generating code coverage report suitable for Coveralls
The Coveralls plugin is hard-coded to only upload the JaCoCo report
associated with the "test" task.  The new "jacocoCoverallsReport" task
creates a single report for both the "test" and "integrationTest" tasks
and stores it in the same location as the report for the "test" task so
it will be picked up by the Coveralls plugin.