Isolate patch computation to worker thread
Changes in the buffer cause a *request* to be made to the worker thread to update the patch.  First, this prevents the EDT from being blocked while waiting for the patch to be computed.  Second, it throttles the rate at which the patch is updated.  That is, instead of updating the patch with every change, the patch is updated as fast as possible.  For example, if three new requests to update the patch are submitted while a patch is actively being computed, when the active patch computation is complete, only a single new patch will be computed which subsumes the three changes that initiated the three requests.

In addition, the worker thread continues to monitor the repository for changes to the HEAD revision of the associated buffer, and forces the patch to be updated if a change is detected.