Allow GitRunner to be created with a program path supplier
The GitPlugin.gitPath() method can change dynamically throughout a jEdit
run.  We've gone to great pains to ensure a new GitRunner is created
every time we need to run Git in order to capture the most recent path
to the Git executable.  This leads to some ugly code.

Now we allow an ISupplier<Path> to be provided for the Git program path
so that a single instance of GitRunner will query for the current Git
program path before each run.  This allows a single GitRunner to be
instantiated once for the lifetime of the buffer.