Modify IGitRunner.run to return an object instead of just the exit code
Also removed the IGitRunner.getProgramPath and getWorkingDirPath
methods.  This was necessary because getProgramPath may have returned a
different path from the one used to actually run Git due to the use of
the supplier object.

Thus, the new GitRunnerResult object returned from run now captures all
the relevant contextual information from the run that may not have been
available to the caller.