Fix potential bugs in BufferAnalyzer integration tests
Files.readAllLines does not provide an explicit final line as jEdit does
via Buffer.  We need to read the entire file and split the lines
ourselves to ensure the final line is explicitly specified.

Spock stubs use memoization.  Therefore, the IBuffer.getLines
implementation will be cached after the first call, and the second and
subsequent calls will return the same value regardless of whether or not
the file content has changed.