Use specific array type instead of generic type
PMD thinks the generic version of this private method is never called,
presumably due to a bug in the algorithm it uses to count call sites
related to generic types.  As this is a temporary method until Eclipse
bug 484897 is fixed, I don't see changing the type from T to String as a
big deal.