Suppress rule violations using annotations instead of comments
I reversed myself on this issue.  The Checkstyle mechanism for
suppressing rule violations via comments is too verbose because it
requires a minimum of two comments (OFF followed by ON) even if only one
line is affected.  The `SuppressWarnings` annotation is cleaner, and it
is now worth it to disable warnings about unknown `SuppressWarnings`
tokens.