Do not use `||` operator in data tables
CodeNarc complains about unnecessary boolean expressions and increased complexity because it doesn't realize this operator has been overridden in a Spock context.  I don't want to disable these rules, so I'm just replacing `||` with `|` and losing the verbosity in the table of separating the input and output parameters.  The generated bytecode is identical.