Preinstrument classes used for acceptance tests
JaCoCo uses a hash of the class at runtime to uniquely identify each class.  The classes used for unit and integration tests have the same identifier because those tests use the class files from the source set output.  However, the acceptance tests use classes from the final distribution.  The classes in the distribution have been modified by the `shadow` plugin because we jarjar the dependencies.  Thus, the identifier for the classes used during acceptance testing differ from those used in unit/integration tests, which prevents us from merging coverage metrics from all three types of tests.

The solution employed in this change is to preinstrument the classes used for acceptance testing before they are jarjar'ed.  This prevents the JaCoCo runtime agent from calculating a different class identifier.  The downside is that we have to create two separate, but nearly identical, distributions: one for production and one for acceptance testing.  In the future, we will add some kind of verification task to ensure the two Jars are identical in all respects except file sizes.  In addition, we may add a separate smoke test task that runs a few acceptance tests directly against the production distribution.