Add smoke tests
The intention of the smoke tests are to minimally test the actual production distribution.  Normally, we would rely on the acceptance tests to validate the distribution, but due to issues with code coverage (see `f4037b7`) the acceptance tests use an instrumented distribution rather than the production version distributed to users.  The smoke tests simply sanity check there is nothing wrong with the production distribution.